+++ CHANGELOG.md
@@ -6,11 +6,18 @@ The project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html
 
 ## [Pending release]
 
+## [5.1.0] - 2021-05-25
+
+### Added
+
+- `binance` BSwap (liquidity pools) endpoints
+- `binance` leveraged token (BLVT) endpoints 
+
 ## [5.0.0] - 2021-05-23
 
 ### Added
 
-- binance COIN-M futures
+- `binance` COIN-M futures
 
 ### Changed
 
@@ -150,7 +157,8 @@ bitvavo.enums.CandelstickInterval -> bitvavo.enums.CandlestickInterval
 
 The official release of `cryptoxlib-aio`.
 
-[Pending release]: https://github.com/nardew/cryptoxlib-aio/compare/5.0.0...HEAD
+[Pending release]: https://github.com/nardew/cryptoxlib-aio/compare/5.1.0...HEAD
+[5.1.0]: https://github.com/nardew/cryptoxlib-aio/compare/5.0.0...5.1.0
 [5.0.0]: https://github.com/nardew/cryptoxlib-aio/compare/4.1.0...5.0.0
 [4.1.0]: https://github.com/nardew/cryptoxlib-aio/compare/4.0.0...4.1.0
 [4.0.0]: https://github.com/nardew/cryptoxlib-aio/compare/3.10.0...4.0.0